Warn without error when processes input is started on Windows
