+++ plugins/inputs/processes/README.md
@@ -6,7 +6,9 @@ them by status (zombie, sleeping, running, etc.)
 On linux this plugin requires access to procfs (/proc), on other OSes
 it requires access to execute `ps`.
 
-### Configuration:
+**Supported Platforms**: Linux, FreeBSD, Darwin
+
+### Configuration
 
 ```toml
 # Get the number of processes and group them by status
@@ -19,9 +21,10 @@ Using the environment variable `HOST_PROC` the plugin will retrieve process info
 
 `docker run -v /proc:/rootfs/proc:ro -e HOST_PROC=/rootfs/proc`
 
-### Measurements & Fields:
+### Metrics
 
 - processes
+  - fields:
     - blocked (aka disk sleep or uninterruptible sleep)
     - running
     - sleeping
@@ -53,14 +56,8 @@ Linux  FreeBSD  Darwin  meaning
   W       W      none   paging (linux kernel < 2.6 only), wait (freebsd)
 ```
 
-### Tags:
-
-None
-
-### Example Output:
+### Example Output
 
 ```
-$ telegraf --config ~/ws/telegraf.conf --input-filter processes --test
-* Plugin: processes, Collection 1
-> processes blocked=8i,running=1i,sleeping=265i,stopped=0i,total=274i,zombie=0i,dead=0i,paging=0i,total_threads=687i 1457478636980905042
+processes blocked=8i,running=1i,sleeping=265i,stopped=0i,total=274i,zombie=0i,dead=0i,paging=0i,total_threads=687i 1457478636980905042
 ```
