+++ plugins/inputs/processes/processes_windows.go
@@ -1,3 +1,27 @@
 // +build windows
 
 package processes
+
+import (
+	"github.com/influxdata/telegraf"
+	"github.com/influxdata/telegraf/plugins/inputs"
+)
+
+type Processes struct {
+	Log telegraf.Logger
+}
+
+func (e *Processes) Init() error {
+	e.Log.Warn("Current platform is not supported")
+	return nil
+}
+
+func (e *Processes) Gather(acc telegraf.Accumulator) error {
+	return nil
+}
+
+func init() {
+	inputs.Add("processes", func() telegraf.Input {
+		return &Processes{}
+	})
+}
